feat: display title in header during form editing
